Unify plugin ui with standard UI; add prepareFeeDisplay and prepareChainIdDisplay
