+++ doc/ethapp_plugins.asc
@@ -3,6 +3,9 @@ Ethereum application Plugins : Technical Specifications
 Ledger Firmware Team <hello@ledger.fr>
 Specification version 1.0 - 24th of September 2020
 
+## 2.0 
+  - Add field in pluginSharedRO, breaking change.
+
 ## 1.0 
   - Initial release
