+++ src_features/signTx/ui_flow_signTx.c
@@ -2,6 +2,80 @@
 #include "ui_callbacks.h"
 #include "chainConfig.h"
 #include "utils.h"
+#include "feature_signTx.h"
+#include "eth_plugin_handler.h"
+
+void plugin_ui_get_id() {
+    ethQueryContractID_t pluginQueryContractID;
+    eth_plugin_prepare_query_contract_ID(&pluginQueryContractID,
+                                         strings.common.fullAddress,
+                                         sizeof(strings.common.fullAddress),
+                                         strings.common.fullAmount,
+                                         sizeof(strings.common.fullAmount));
+    // Query the original contract for ID if it's not an internal alias
+    if (!eth_plugin_call(ETH_PLUGIN_QUERY_CONTRACT_ID, (void *) &pluginQueryContractID)) {
+        PRINTF("Plugin query contract ID call failed\n");
+        io_seproxyhal_touch_tx_cancel(NULL);
+    }
+}
+
+void plugin_ui_get_item() {
+    ethQueryContractUI_t pluginQueryContractUI;
+    eth_plugin_prepare_query_contract_UI(&pluginQueryContractUI,
+                                         dataContext.tokenContext.pluginUiCurrentItem,
+                                         strings.common.fullAddress,
+                                         sizeof(strings.common.fullAddress),
+                                         strings.common.fullAmount,
+                                         sizeof(strings.common.fullAmount));
+    if (!eth_plugin_call(ETH_PLUGIN_QUERY_CONTRACT_UI, (void *) &pluginQueryContractUI)) {
+        PRINTF("Plugin query contract UI call failed\n");
+        io_seproxyhal_touch_tx_cancel(NULL);
+    }
+}
+
+void display_next_plugin_item(bool entering) {
+    if (entering) {
+        if (dataContext.tokenContext.pluginUiState == PLUGIN_UI_OUTSIDE) {
+            dataContext.tokenContext.pluginUiState = PLUGIN_UI_INSIDE;
+            dataContext.tokenContext.pluginUiCurrentItem = 0;
+            plugin_ui_get_item();
+            ux_flow_next();
+        } else {
+            if (dataContext.tokenContext.pluginUiCurrentItem > 0) {
+                dataContext.tokenContext.pluginUiCurrentItem--;
+                plugin_ui_get_item();
+                ux_flow_next();
+            } else {
+                dataContext.tokenContext.pluginUiState = PLUGIN_UI_OUTSIDE;
+                dataContext.tokenContext.pluginUiCurrentItem = 0;
+                ux_flow_prev();
+            }
+        }
+    } else {
+        if (dataContext.tokenContext.pluginUiState == PLUGIN_UI_OUTSIDE) {
+            dataContext.tokenContext.pluginUiState = PLUGIN_UI_INSIDE;
+            plugin_ui_get_item();
+            ux_flow_prev();
+        } else {
+            if (dataContext.tokenContext.pluginUiCurrentItem <
+                dataContext.tokenContext.pluginUiMaxItems - 1) {
+                dataContext.tokenContext.pluginUiCurrentItem++;
+                plugin_ui_get_item();
+                ux_flow_prev();
+                // Reset multi page layout to the first page
+                G_ux.layout_paging.current = 0;
+#ifdef TARGET_NANOS
+                ux_layout_paging_redisplay_by_addr(G_ux.stack_count - 1);
+#else
+                ux_layout_bnnn_paging_redisplay(0);
+#endif
+            } else {
+                dataContext.tokenContext.pluginUiState = PLUGIN_UI_OUTSIDE;
+                ux_flow_next();
+            }
+        }
+    }
+}
 
 // clang-format off
 UX_STEP_NOCB(
@@ -108,6 +182,40 @@ UX_STEP_NOCB(
       .title = "Address",
       .text = strings.common.fullAddress,
     });
+
+UX_STEP_NOCB_INIT(
+  ux_plugin_approval_id_step,
+  bnnn_paging,
+  plugin_ui_get_id(),
+  {
+    .title = strings.common.fullAddress,
+    .text = strings.common.fullAmount
+  });
+
+UX_STEP_INIT(
+  ux_plugin_approval_before_step,
+  NULL,
+  NULL,
+  {
+    display_next_plugin_item(true);
+  });
+
+UX_FLOW_DEF_NOCB(
+  ux_plugin_approval_display_step,
+  bnnn_paging,
+  {
+    .title = strings.common.fullAddress,
+    .text = strings.common.fullAmount
+  });
+
+UX_STEP_INIT(
+  ux_plugin_approval_after_step,
+  NULL,
+  NULL,
+  {
+    display_next_plugin_item(false);
+  });
+
 UX_STEP_NOCB(
     ux_approval_base_fee_step,
     bnnn_paging,
@@ -136,6 +244,7 @@ UX_STEP_NOCB(
       .title = "Chain ID",
       .text = strings.common.chainID,
     });
+
 UX_STEP_CB(
     ux_approval_accept_step,
     pbb,
@@ -171,16 +280,28 @@ UX_STEP_NOCB(ux_approval_data_warning_step,
     });
 // clang-format on
 
-const ux_flow_step_t *ux_approval_tx_flow[12];
+const ux_flow_step_t *ux_approval_tx_flow[15];
 
-void ux_approve_tx(bool dataPresent) {
+void ux_approve_tx(bool fromPlugin) {
     int step = 0;
     ux_approval_tx_flow[step++] = &ux_approval_review_step;
+
     if (tmpContent.txContent.dataPresent && !N_storage.contractDetails) {
         ux_approval_tx_flow[step++] = &ux_approval_data_warning_step;
     }
-    ux_approval_tx_flow[step++] = &ux_approval_amount_step;
-    ux_approval_tx_flow[step++] = &ux_approval_address_step;
+
+    if (fromPlugin) {
+        prepareChainIdDisplay();
+        prepareFeeDisplay();
+        ux_approval_tx_flow[step++] = &ux_plugin_approval_id_step;
+        ux_approval_tx_flow[step++] = &ux_plugin_approval_before_step;
+        ux_approval_tx_flow[step++] = &ux_plugin_approval_display_step;
+        ux_approval_tx_flow[step++] = &ux_plugin_approval_after_step;
+    } else {
+        ux_approval_tx_flow[step++] = &ux_approval_amount_step;
+        ux_approval_tx_flow[step++] = &ux_approval_address_step;
+    }
+
     if (N_storage.displayNonce) {
         ux_approval_tx_flow[step++] = &ux_approval_nonce_step;
     }
